Link Client to User object for authenticated users
This only applies to new item reservations.
No change is needed for loans since it transparently uses Client.

I've manually tested this by reserving an item when logged in and when logged out
to ensure the expected data is stored. I'll add automated tests after the
conflicting test changes in #171 are merged.