Replace Client foreign keys with User foreign keys
All ItemLoans and ItemReservations previously had a Client as a foreign
key, but Client is an unnecessary abstraction that was created before.
It is even more unnecessary now that all "clients" need to log in prior
to reserving an item.

The database migration included performs a completely automated
transition, but note that it cannot necessarily be rolled back since the
Client model (and hence all those objects) will be deleted.